streams-processing/requirements.txt
fastapi
uvicorn
kafka-python
faker
pydantic
python-dateutil>=2.8.2
requests>=2.28.1